fix: correct pagination overflow and literal search in listings
Two defects found by probing the running API against PostgreSQL.

Offset overflow: `(page - 1) * pageSize` was computed in int, and `page` is
only bounded from below, so `?page=2147483647&pageSize=100` wrapped into a
negative OFFSET, which PostgreSQL rejects. Every listing answered 500 instead
of an empty page. The multiplication is now widened to long.

Unescaped LIKE wildcards: the network filter interpolated user input straight
into an ILIKE pattern, so '%' and '_' behaved as wildcards. Searching for '_'
returned every network, and a name containing those characters could not be
searched literally. They are now escaped, with an explicit ESCAPE clause.

Both are covered by integration tests that were confirmed to fail before the
fix and pass after it. The coverage floor moves to 70% to match the 74.9% the
new tests reach.

Also aligns the README with the code: /transfers/high-energy filters with >=,
so it lists transfers from the threshold, not strictly above it.

+++ EcoMyceliumTracker/Repositories/SensorRepository.cs
@@ -28,7 +28,9 @@ public sealed class SensorRepository(NpgsqlDataSource dataSource) : ISensorRepos
             NetworkId = networkId,
             IsActive = isActive,
             PageSize = pageSize,
-            Offset = (page - 1) * pageSize
+            // Widened before multiplying: page is only bounded from below, so
+            // int arithmetic here overflows into a negative OFFSET.
+            Offset = (long)(page - 1) * pageSize
         };
         var sql = $$"""
             SELECT {{SelectColumns}}
